Remove abstractions
- Remove abstractions now there is only one processing implementation.
- Move "job" implementation back into `GitHubWebhookService`.
