--- src/Costellobot/IGitHubEventHandler.cs
@@ -1,9 +1,0 @@
-﻿// Copyright (c) Martin Costello, 2022. All rights reserved.
-// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
-
-namespace MartinCostello.Costellobot;
-
-public interface IGitHubEventHandler
-{
-    Task HandleAsync(GitHubEvent payload, CancellationToken cancellationToken);
-}
